+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Adding CHANGELOG.md
 - Removing homepage from gemspec as we don't have proper one yet
 - Adding source code URL
+- Allowing to connect to Redis with no sentinels
 
 ## 1.0.0
 - Initial commit!